format files

+++ contracts/Marketplace/Marketplace.sol
@@ -10,7 +10,6 @@ import "../PaymentRouter/IPaymentRouter.sol";
 import "../Tokens/IPazariTokenMVP.sol";
 
 contract AccessControlMP {
-
   // Maps admin addresses to bool
   mapping(address => bool) public isAdmin;
 
@@ -28,15 +27,27 @@ contract AccessControlMP {
   event AdminRemoved(address oldAdmin, address adminAuthorized, string memo, uint256 timestamp);
 
   // Fires when item admins are added or removed
-  event ItemAdminAdded(uint256 itemID, address newAdmin, address adminAuthorized, string memo, uint256 timestamp);
-  event ItemAdminRemoved(uint256 itemID, address oldAdmin, address adminAuthorized, string memo, uint256 timestamp);
+  event ItemAdminAdded(
+    uint256 itemID,
+    address newAdmin,
+    address adminAuthorized,
+    string memo,
+    uint256 timestamp
+  );
+  event ItemAdminRemoved(
+    uint256 itemID,
+    address oldAdmin,
+    address adminAuthorized,
+    string memo,
+    uint256 timestamp
+  );
 
   // Fires when an address is blacklisted/whitelisted from the Pazari Marketplace
   event AddressBlacklisted(address blacklistedAddress, address adminAddress, string memo, uint256 timestamp);
   event AddressWhitelisted(address whitelistedAddress, address adminAddress, string memo, uint256 timestamp);
 
   constructor(address[] memory _adminAddresses) {
-    for(uint i = 0; i < _adminAddresses.length; i++){
+    for (uint256 i = 0; i < _adminAddresses.length; i++) {
       isAdmin[_adminAddresses[i]] = true;
     }
   }
@@ -46,69 +57,75 @@ contract AccessControlMP {
    * for everything else. See PaymentRouter for more details.
    */
   function _msgSender() public view returns (address) {
-    if(tx.origin != msg.sender && isAdmin[msg.sender]){
+    if (tx.origin != msg.sender && isAdmin[msg.sender]) {
       return tx.origin;
-    }
-    else return msg.sender;
+    } else return msg.sender;
   }
-  
+
   // Adds an address to isAdmin mapping
   function addAdmin(address _newAddress, string memory _memo) external onlyAdmin returns (bool) {
     require(!isAdmin[_newAddress], "Address is already an admin");
-    
+
     isAdmin[_newAddress] = true;
-    
+
     emit AdminAdded(_newAddress, tx.origin, _memo, block.timestamp);
     return true;
   }
 
   // Adds an address to isItemAdmin mapping
-  function addItemAdmin(uint256 _itemID, address _newAddress, string memory _memo) external onlyItemAdmin(_itemID) returns (bool) {
-    require(isItemAdmin[_itemID][msg.sender] && isItemAdmin[_itemID][tx.origin], "Caller is not admin");    
-    require(!isItemAdmin[_itemID][_newAddress], "Address is already an item admin");    
+  function addItemAdmin(
+    uint256 _itemID,
+    address _newAddress,
+    string memory _memo
+  ) external onlyItemAdmin(_itemID) returns (bool) {
+    require(isItemAdmin[_itemID][msg.sender] && isItemAdmin[_itemID][tx.origin], "Caller is not admin");
+    require(!isItemAdmin[_itemID][_newAddress], "Address is already an item admin");
 
     isItemAdmin[_itemID][_newAddress] = true;
 
     emit ItemAdminAdded(_itemID, _newAddress, _msgSender(), _memo, block.timestamp);
-    return true;   
+    return true;
   }
 
   // Removes an address from isAdmin mapping
   function removeAdmin(address _oldAddress, string memory _memo) external onlyAdmin returns (bool) {
     require(isAdmin[_oldAddress], "Address is not an admin");
-    
+
     isAdmin[_oldAddress] = false;
-    
+
     emit AdminRemoved(_oldAddress, tx.origin, _memo, block.timestamp);
     return true;
   }
 
   // Removes an address from isItemAdmin mapping
-  function removeItemAdmin(uint256 _itemID, address _oldAddress, string memory _memo) external onlyItemAdmin(_itemID) returns (bool) {
-    require(isItemAdmin[_itemID][msg.sender] && isItemAdmin[_itemID][tx.origin], "Caller is not admin");    
-    require(isItemAdmin[_itemID][_oldAddress], "Address is not an admin");    
+  function removeItemAdmin(
+    uint256 _itemID,
+    address _oldAddress,
+    string memory _memo
+  ) external onlyItemAdmin(_itemID) returns (bool) {
+    require(isItemAdmin[_itemID][msg.sender] && isItemAdmin[_itemID][tx.origin], "Caller is not admin");
+    require(isItemAdmin[_itemID][_oldAddress], "Address is not an admin");
     require(itemCreator[_itemID] == _msgSender(), "Cannot remove item creator");
 
     isItemAdmin[_itemID][_oldAddress] = false;
 
     emit ItemAdminRemoved(_itemID, _oldAddress, _msgSender(), _memo, block.timestamp);
     return true;
-  }  
-  
+  }
+
   /**
    * @notice Toggles isBlacklisted for an address. Can only be called by Pazari
    * Marketplace admins. Other contracts that implement address blacklisting
    * can call this contract's isBlacklisted mapping.
    */
-  function toggleBlacklist(address _listedAddress, string memory _memo) external returns (bool){
+  function toggleBlacklist(address _listedAddress, string memory _memo) external returns (bool) {
     require(isAdmin[msg.sender] && isAdmin[tx.origin], "Only Pazari admin");
     require(!isAdmin[_listedAddress], "Cannot blacklist admins");
 
-    if(!isBlacklisted[_listedAddress]){
+    if (!isBlacklisted[_listedAddress]) {
       isBlacklisted[_listedAddress] = true;
       emit AddressBlacklisted(_listedAddress, _msgSender(), _memo, block.timestamp);
-    }
-    else {
+    } else {
       isBlacklisted[_listedAddress] = false;
       emit AddressWhitelisted(_listedAddress, _msgSender(), _memo, block.timestamp);
     }
@@ -116,7 +133,6 @@ contract AccessControlMP {
     return true;
   }
 
- 
   /**
    * @notice Requires that both msg.sender and tx.origin be admins. This restricts all
    * calls to only Pazari-owned admin addresses, including wallets and contracts, and
@@ -125,7 +141,7 @@ contract AccessControlMP {
   modifier onlyAdmin() {
     require(isAdmin[msg.sender] && isAdmin[tx.origin], "Only Pazari-owned addresses");
     _;
-  } 
+  }
 
   modifier noBlacklist() {
     require(!isBlacklisted[_msgSender()], "Caller cannot be blacklisted");
@@ -135,14 +151,11 @@ contract AccessControlMP {
   // Restricts access to admins of a MarketItem
   modifier onlyItemAdmin(uint256 _itemID) {
     require(
-      itemCreator[_itemID] == _msgSender() || 
-      isItemAdmin[_itemID][_msgSender()] ||
-      isAdmin[_msgSender()],
+      itemCreator[_itemID] == _msgSender() || isItemAdmin[_itemID][_msgSender()] || isAdmin[_msgSender()],
       "Caller is neither admin nor item creator"
     );
     _;
   }
-
 }
 
 contract Marketplace is ERC1155Holder, AccessControlMP {
@@ -207,7 +220,7 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
     bool forSale;
     uint256 itemLimit;
   }
-  
+
   // Counter for items with forSale == false
   Counters.Counter private itemsSoldOut;
 
@@ -342,7 +355,7 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
     // Check that Marketplace's internal balance matches the token's balanceOf() value
     MarketItem memory item = marketItems[itemID - 1];
     assert(IERC1155(item.tokenContract).balanceOf(address(this), item.tokenID) == item.amount);
-  }  
+  }
 
   /**
    * @dev Private function that updates internal variables and storage for a new MarketItem
@@ -392,7 +405,7 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
     itemCreator[itemID] = _msgSender();
     // Pushes MarketItem to marketItems[]
     marketItems.push(item);
-  
+
     // Push itemID to sellersMarketItems mapping array
     // _msgSender == sellerAddress
     sellersMarketItems[_msgSender()].push(item.itemID);
@@ -402,15 +415,7 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
 
     // Emits MarketItemCreated event
     // _msgSender == sellerAddress
-    emit MarketItemCreated(
-      itemID,
-      _tokenContract,
-      _tokenID,
-      _msgSender(),
-      _price,
-      _amount,
-      _paymentContract
-    );
+    emit MarketItemCreated(itemID, _tokenContract, _tokenID, _msgSender(), _price, _amount, _paymentContract);
   }
 
   /**
@@ -439,7 +444,7 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
     require(balance < itemLimit, "Buyer already owns the item limit");
     // If _amount == 0, then purchase itemLimit - balance
     // If _amount + balance surpasses itemLimit, then purchase itemLimit - balance
-    if(_amount == 0 || _amount + balance > itemLimit){
+    if (_amount == 0 || _amount + balance > itemLimit) {
       _amount = itemLimit - balance;
     }
 
@@ -461,10 +466,16 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
     require(IERC20(item.paymentContract).approve(address(this), totalCost), "ERC20 approval failure");
 
     // Pull payment tokens from msg.sender to Marketplace
-    require(IERC20(item.paymentContract).transferFrom(_msgSender(), address(this), totalCost), "ERC20 transfer failure");
+    require(
+      IERC20(item.paymentContract).transferFrom(_msgSender(), address(this), totalCost),
+      "ERC20 transfer failure"
+    );
 
     // Approve payment tokens for transfer to PaymentRouter
-    require(IERC20(item.paymentContract).approve(address(iPaymentRouter), totalCost), "ERC20 approval failure");
+    require(
+      IERC20(item.paymentContract).approve(address(iPaymentRouter), totalCost),
+      "ERC20 approval failure"
+    );
 
     // Send ERC20 tokens through PaymentRouter, isPush determines which function is used
     // note PaymentRouter functions make external calls to ERC20 contracts, thus they are interactions
@@ -488,11 +499,19 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
    *
    * @dev Emits ItemRestocked event
    */
-  function restockItem(uint256 _itemID, uint256 _amount) external noBlacklist onlyItemAdmin(_itemID) returns (bool) {
+  function restockItem(uint256 _itemID, uint256 _amount)
+    external
+    noBlacklist
+    onlyItemAdmin(_itemID)
+    returns (bool)
+  {
     /* ========== CHECKS ========== */
     MarketItem memory item = marketItems[_itemID - 1];
     require(marketItems.length >= _itemID, "MarketItem does not exist");
-    require(IERC1155(item.tokenContract).balanceOf(_msgSender(), item.tokenID) >= _amount, "Insufficient token balance");
+    require(
+      IERC1155(item.tokenContract).balanceOf(_msgSender(), item.tokenID) >= _amount,
+      "Insufficient token balance"
+    );
 
     /* ========== EFFECTS ========== */
     marketItems[_itemID - 1].amount += _amount;
@@ -501,7 +520,9 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
     /* ========== INTERACTIONS ========== */
     IERC1155(item.tokenContract).safeTransferFrom(_msgSender(), address(this), item.tokenID, _amount, "");
 
-    assert(IERC1155(item.tokenContract).balanceOf(address(this), item.tokenID) == marketItems[_itemID - 1].amount);
+    assert(
+      IERC1155(item.tokenContract).balanceOf(address(this), item.tokenID) == marketItems[_itemID - 1].amount
+    );
     return true;
   }
 
@@ -513,7 +534,12 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
    *
    * @dev Emits StockPulled event
    */
-  function pullStock(uint256 _itemID, uint256 _amount) external noBlacklist onlyItemAdmin(_itemID) returns (bool) {
+  function pullStock(uint256 _itemID, uint256 _amount)
+    external
+    noBlacklist
+    onlyItemAdmin(_itemID)
+    returns (bool)
+  {
     /* ========== CHECKS ========== */
     // itemID will always be <= marketItems.length, but cannot be > marketItems.length
     require(_itemID <= marketItems.length, "MarketItem does not exist");
@@ -720,7 +746,7 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
    * @notice Returns an array of MarketItems created the seller's address
    * @dev No restrictions for calling this
    */
-  function getSellersMarketItems(address _sellerAddress) public view returns (uint256[] memory){
+  function getSellersMarketItems(address _sellerAddress) public view returns (uint256[] memory) {
     return sellersMarketItems[_sellerAddress];
   }
